+++ src/types.js
@@ -4,39 +4,60 @@
  * @property {HTMLElement|SVGElement|HTMLDocument} [area=document] area in which you can drag. If not provided it will be the whole document
  * @property {number} [autoScrollSpeed=1] Speed in which the area scrolls while selecting (if available). Unit is pixel per movement.
  * @property {number} [zoom=1] Zoom scale factor (in case of using CSS style transform: scale() which messes with real positions). Unit scale zoom.
- * @property {DSCallback} [callback=() => {}] a callback function that gets fired when the element is dropped. This callback gets a property which is an array that holds all selected nodes. The second property passed is the event object.
  * @property {boolean} [customStyles=false] if set to true, no styles (except for position absolute) will be applied by default
  * @property {string} [hoverClass=ds-hover] the class assigned to the mouse hovered items
  * @property {boolean} [multiSelectMode=false] Add newly selected elements to the selection instead of replacing them.
- * @property {DSGenericInteractionCallbackEvent} [onDragMove=()=>{}] It is fired when the user drags. This callback gets the event object. Executed before DragSelect function code ran, after getting the current mouse position.
- * @property {DSGenericInteractionCallbackEvent} [onDragStartBegin=()=>{}] Is fired when the user clicks in the area. This callback gets the event object. Executed *before* DragSelect function code ran.
- * @property {DSGenericInteractionCallbackEvent} [onDragStart=()=>{}] It is fired when the user clicks in the area. This callback gets the event object. Executed after DragSelect function code ran, before the setup of event listeners.
- * @property {DSInteractionCallbackEvent} [onElementSelect=()=>{}] It is fired every time an element is selected. This callback gets a property which is the just selected node
- * @property {DSInteractionCallbackEvent} [onElementUnselect=()=>{}] It is fired every time an element is de-selected. This callback gets a property which is the just de-selected node
  * @property {string} [selectableClass=ds-selectable] the class assigned to the elements that can be selected
  * @property {HTMLElement[] | SVGElement[] | HTMLElement | SVGElement} [selectables=[]] the elements that can be selected
  * @property {string} [selectedClass=ds-selected] the class assigned to the selected items
  * @property {HTMLElement} [selector=HTMLElement] the square that will draw the selection
  * @property {string} [selectorClass=ds-selector] the class assigned to the square selector helper
  * @property {string} [selectorAreaClass=ds-selector-area] the class assigned to the square in which the selector resides. By default it's invisible
+ * @property {DSCallbackEvent} [callback] Deprecated: please use DragSelect.subscribe('callback', callback) instead
+ * @property {DSDragMoveEvent} [onDragMove] Deprecated: please use DragSelect.subscribe('onDragMove', onDragMove) instead
+ * @property {DSDragMoveBeginEvent} [onDragStartBegin]  Deprecated: please use DragSelect.subscribe('onDragStartBegin', onDragStartBegin) instead
+ * @property {DSDragStartEvent} [onDragStart]  Deprecated: please use DragSelect.subscribe('onDragStart', onDragStart) instead
+ * @property {DSElementSelectEvent} [onElementSelect]  Deprecated: please use DragSelect.subscribe('onElementSelect', onElementSelect) instead
+ * @property {DSElementUnSelectEvent} [onElementUnselect]  Deprecated: please use DragSelect.subscribe('onElementUnselect', onElementUnselect) instead
  * @property {Array.<'ctrlKey'|'shiftKey'|'metaKey'>} [multiSelectKeys=['ctrlKey', 'shiftKey', 'metaKey']] An array of keys that allows switching to the multi-select mode (see the @multiSelectMode option). The only possible values are keys that are provided via the event object. So far: <kbd>ctrlKey</kbd>, <kbd>shiftKey</kbd>, <kbd>metaKey</kbd> and <kbd>altKey</kbd>. Provide an empty array `[]` if you want to turn off the functionality.
+ */
+
+/**
+ * The Object that is passed back to any callback method
+ * @typedef {Object} CallbackObject
+ * @property {Array<HTMLElement|SVGElement|any>} items The items currently selected
+ * @property {MouseEvent|TouchEvent|Event} [event] The respective event object
+ * @property {HTMLElement|SVGElement|any} [item] The single item currently interacted with
  */
 /**
  * @typedef {function} DSCallback
- * @param {DSElements} selected - The selected items
- * @param {DSEvent} [event]
- * @return {*}
+ * @param {CallbackObject} selected
  */
 /**
- * @typedef {function} DSGenericInteractionCallbackEvent
- * @param {DSEvent} [event]
- * @return {*}
+ * Callback function that gets fired when the element is selected.
+ * @typedef {DSCallback} DSCallbackEvent
  */
 /**
- * @typedef {function} DSInteractionCallbackEvent
- * @param {DSElement} item
- * @return {*}
+ * Fired while the user drags.
+ * @typedef {DSCallback} DSDragMoveEvent
  */
+/**
+ * Fired while the user drags.
+ * @typedef {DSCallback} DSDragMoveBeginEvent
+ */
+/**
+ * When the user clicks in the area.
+ * @typedef {DSCallback} DSDragStartEvent
+ */
+/**
+ * Fired every time an element is selected.
+ * @typedef {DSCallback} DSElementSelectEvent
+ */
+/**
+ * Fired every time an element is un-selected.
+ * @typedef {DSCallback} DSElementUnSelectEvent
+ */
+
 /** @typedef {HTMLElement|SVGElement|HTMLDocument} DSArea area in which you can drag */
 /** @typedef {HTMLElement} DSSelectorArea area in which you can drag */
 /** @typedef {Array.<HTMLElement|SVGElement> | HTMLElement | SVGElement} DSInputElements the elements that can be selected */
@@ -46,3 +67,4 @@
 /** @typedef {number} DSZoom Zoom scale factor. Unit scale zoom */
 /** @typedef {Array.<'ctrlKey'|'shiftKey'|'metaKey'>} DSMultiSelectKeys An array of keys that allows switching to the multi-select mode */
 
+/** @typedef {'dragmove'|'dragstartbegin'|'dragstart'|'elementselect'|'elementunselect'|'callback'} DSCallbackNames the name of the callback */